Todo App

+++ caseStudy/main.js
@@ -1,57 +1,118 @@
 let toDoLists = JSON.parse(localStorage.getItem("toDoLists")) || [];
 let input = document.querySelector('.input');
 let btnAdd = document.querySelector('.btnAdd');
-let lists = document.querySelectorAll('.list');
 let btnRemoveAll = document.querySelector('.btnRemoveAll');
 let infors = document.querySelectorAll('.info');
-let sortAZ = document.querySelector('.sortaz');
-let sortZA = document.querySelector('.sortza');
+let btnSortAZ = document.querySelector('.sortaz');
+let btnSortZA = document.querySelector('.sortza');
+let list = document.querySelector('.list');
 let save;
+let STATUS = ['pending', 'doing', 'done'];
 
 function showToDo(status) {
     if (toDoLists) {
         let statusList = status === 'all' ? toDoLists : toDoLists.filter(item => item.status === status);
         let htmls = statusList.map(function (item, index) {
             return `
-                      <li>
-                          <span class="list-title ${item.status}" onclick='handleComplete(${JSON.stringify(item)}, ${index})'>${item.name}</span>
+                      <li class="list-item">
+                          <span class="list-title ${item.status}">${item.name}</span>
                           <div>
-                                ${item.status === 'completed' ? '' :
-                                    `<button class="btn btnEdit" onclick='editToDo(${JSON.stringify(item)}, ${index})'>
-                                    <i class="fa-solid fa-pen-to-square"></i>
-                                    </button>`
+                          <button class="btnSetting" onclick='showSetting(this, ${index})'><i class="fa-solid fa-gear"></i></button>
+                          <div class="setting">
+                                ${item.status === 'done' ? '' :
+                                    `<div class="option option-edit" onclick='editToDo(${JSON.stringify(item)}, ${index})'>
+                                        <i class="fa-solid fa-pen-to-square"></i> 
+                                        <span>Edit</span>
+                                    </div>`
                                 }
-                                    <button class="btn btnDelete" onclick="deleteTodo(${index})"><i class="fa-solid fa-trash"></i></button>
+                                    <div class="option option-delete" onclick="deleteTodo(${index})">
+                                        <i class="fa-regular fa-trash-can"></i> 
+                                        <span>Delete</span>
+                                    </div>
+                                    <div class="option" onclick='setStatus(${JSON.stringify(item)}, ${index}, "pending")'>
+                                        <i class="fa-solid fa-plus"></i>
+                                        <span>Add to pending</span>
+                                    </div>
+                                    <div class="option" onclick='setStatus(${JSON.stringify(item)}, ${index}, "doing")'>
+                                        <i class="fa-solid fa-plus"></i>
+                                        <span>Add to doing</span>
+                                    </div>
+                                    <div class="option" onclick='setStatus(${JSON.stringify(item)}, ${index}, "done")'>
+                                        <i class="fa-solid fa-plus"></i>
+                                        <span>Add to done</span>
+                                    </div>
+                          </div>
                           </div>
                        </li>
                     `
         })
-        document.querySelector(`.list.${status}`).innerHTML = htmls.join('');
+        list.innerHTML = htmls.join('');
+        console.log(window.innerHeight * 0.5)
+        list.scrollHeight > window.innerHeight * 0.5 ? list.classList.add('overflow') : list.classList.remove('overflow');
+    }
+}
+
+function showNotice() {
+    document.querySelector(`.notice.all`).innerText = toDoLists.length;
+    for(let i of STATUS){
+        document.querySelector(`.notice.${i}`).innerText = toDoLists.filter(item => item.status === i).length;
     }
 }
 function render() {
-    showToDo('all');
-    showToDo('pending');
-    showToDo('completed');
+    let status = document.querySelector('.info.active').id;
+    showToDo(status);
+    showNotice();
 }
 
 window.onload = render;
+
+function showSetting(element, index){
+    let setting = document.querySelectorAll('.setting')[index];
+    if(document.querySelector('.setting.show') && !setting.classList.contains('show')) {
+        document.querySelector('.setting.show').classList.remove('show');
+    }
+    setting.classList.toggle('show');
+
+    function removeShow (e){
+        if(e.target !== element && e.target.tagName !== 'I') {
+            setting.classList.remove('show');
+            document.removeEventListener('click', removeShow);
+        }
+    }
+    document.addEventListener('click', removeShow);
+}
+
+function setStatus(item, index, status){
+    item.status = status;
+    toDoLists[index] = item;
+    localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
+    render()
+}
 function addToDo() {
     if (btnAdd.innerText === 'Save') {
-        toDoLists[save].name = input.value.trim()[0].toUpperCase() + input.value.trim().slice(1);
+        let str = input.value.trim()[0].toUpperCase() + input.value.trim().slice(1);
+        let index = toDoLists.findIndex(todo => todo.name === str);
+        if(index === -1) {
+            toDoLists[save].name = str;
+            alert('Success')
+        } else alert('Error')
         btnAdd.innerText = 'Add';
     } else {
         if (input.value.trim()) {
             let str = input.value.trim()[0].toUpperCase() + input.value.trim().slice(1);
-            let newToDo = { 'name': str, 'status': 'pending' };
-            toDoLists.push(newToDo);
+            let index = toDoLists.findIndex(todo => todo.name === str);
+            if(index === -1) {
+                let newToDo = {'name': str, 'status': 'pending'};
+                toDoLists.push(newToDo);
+                alert('Success')
+            } else alert('Error')
         }
     }
     localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
 
-    if (sortAZ.classList.contains('active')) {
+    if (btnSortAZ.classList.contains('active')) {
         toDoLists.sort((a, b) => a.name.localeCompare(b.name));
-    } else if (sortZA.classList.contains('active')) {
+    } else if (btnSortZA.classList.contains('active')) {
         toDoLists.sort((a, b) => a.name.localeCompare(b.name)).reverse();
     }
 
@@ -60,9 +121,12 @@ function addToDo() {
     render();
 }
 function deleteTodo(index) {
-    toDoLists.splice(index, 1);
-    localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
-    render();
+    let isConfirm = confirm('Xác nhận xóa');
+    if(isConfirm){
+        toDoLists.splice(index, 1);
+        localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
+        render();
+    }
 }
 
 function editToDo(item, index) {
@@ -72,55 +136,53 @@ function editToDo(item, index) {
     save = index;
 }
 
-function handleComplete(item, index) {
-    item.status = item.status === 'pending' ? 'completed' : 'pending';
-    toDoLists[index] = item;
-    localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
-    render();
+function removeAll(){
+    let status = document.querySelector('.info.active').id;
+    let isConfirm = confirm(`Xác nhận xóa tất cả trong mục ${status[0].toUpperCase() + status.slice(1)}`);
+    if(isConfirm) {
+        if (status === 'all') {
+            toDoLists = [];
+        } else {
+            toDoLists = toDoLists.filter(todo => todo.status !== status);
+        }
+        localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
+        render();
+    }
 }
 
 btnAdd.addEventListener('click', addToDo);
-
 input.addEventListener('keydown', function (e) {
     if (e.key === 'Enter') addToDo();
 })
+btnRemoveAll.addEventListener('click', removeAll);
 
-btnRemoveAll.addEventListener('click', function () {
-    toDoLists = []
-    localStorage.setItem("toDoLists", JSON.stringify(toDoLists));
-    render();
-})
-
-infors.forEach((info, index) => {
+infors.forEach((info) => {
     info.onclick = function () {
-        infors.forEach(item => item.classList.remove('active'));
+        document.querySelector('.info.active').classList.remove('active');
         this.classList.add('active');
-
-        lists.forEach(item => item.classList.remove('show'));
-        lists[index].classList.add('show');
+        render();
     }
 })
 
-sortAZ.onclick = function () {
-    if (this.classList.contains('active')) {
-        this.classList.remove('active');
-        toDoLists = JSON.parse(localStorage.getItem("toDoLists"))
-    } else {
-        sortZA.classList.remove('active');
-        this.classList.add('active');
-        toDoLists.sort((a, b) => a.name.localeCompare(b.name));
-    }
-    render();
-}
-
-sortZA.onclick = function () {
-    if (this.classList.contains('active')) {
-        this.classList.remove('active');
+function sort (btn1, btn2,order){
+    if (btn1.classList.contains('active')) {
+        btn1.classList.remove('active');
         toDoLists = JSON.parse(localStorage.getItem("toDoLists"));
     } else {
-        sortAZ.classList.remove('active');
-        this.classList.add('active');
-        toDoLists.sort((a, b) => a.name.localeCompare(b.name)).reverse();
+        btn2.classList.remove('active');
+        btn1.classList.add('active');
+        if(order === 'ASC')
+            toDoLists.sort((a, b) => a.name.localeCompare(b.name));
+        else
+            toDoLists.sort((a, b) => b.name.localeCompare(a.name));
     }
     render();
 }
+
+btnSortAZ.addEventListener('click', function (){
+    sort(btnSortAZ, btnSortZA, 'ASC');
+})
+
+btnSortZA.addEventListener('click', function (){
+    sort(btnSortZA, btnSortAZ, 'DSC');
+})
